New System Messages started

+++ app/src/main/java/com/Alex/diary/XLSParser.java
@@ -1,5 +1,7 @@
 package com.Alex.diary;
 
+import static com.Alex.diary.ui.all_marks.MarksAllFragment.Load;
+
 import android.annotation.SuppressLint;
 import android.app.DownloadManager;
 import android.content.BroadcastReceiver;
@@ -8,15 +10,21 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.database.Cursor;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
 import android.util.Log;
 import android.webkit.WebView;
-import android.widget.ListView;
 import android.widget.Toast;
 
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.ActivityCompat;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentTransaction;
+
+import com.Alex.diary.ui.all_marks.MarksAllFragment;
+import com.Alex.diary.ui.error.ErrorFragment;
+import com.Alex.diary.ui.full_info_marks.MarksInfoFragment;
 
 import org.apache.poi.hssf.usermodel.HSSFDateUtil;
 import org.apache.poi.hssf.usermodel.HSSFWorkbook;
@@ -29,7 +37,6 @@ import org.apache.poi.ss.usermodel.Workbook;
 
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
@@ -40,16 +47,29 @@ public class XLSParser extends AppCompatActivity {
     private WebView webView;
     private String url;
     private String cookies;
+    public static Context contextt;
+    public File downloaded;
+    public static FragmentManager fm;
     String destinationName = "MarksQuarter.xls";
-    public List<List> Items;
+    public static List<List> Items;
     File file = null;
-
+    public static boolean IsMarksInfo = false;
+    public static int PositionSelected;
+    static MarksInfoFragment marksInfoFragment;
+    static MarksAllFragment marksAllFragment;
     @SuppressLint("SetJavaScriptEnabled")
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_xls);
-        checkDownloadPermission();
+        //checkDownloadPermission();
+        contextt = getApplicationContext();
+        fm = getSupportFragmentManager();
+        FragmentTransaction ft = fm.beginTransaction();
+        marksInfoFragment = new MarksInfoFragment();
+        marksAllFragment = new MarksAllFragment();
+        ft.replace(R.id.fragmentXLS, marksAllFragment);
+        ft.commit();
         DeleteOld(destinationName);
         cookies = getIntent().getStringExtra("gfjxsasd");
         url = getIntent().getStringExtra("djhjgfj");
@@ -69,18 +89,17 @@ public class XLSParser extends AppCompatActivity {
 
 
     }
-    private void checkDownloadPermission() {
+    /**
+     * Check if the user has already given permission to download files. If not, ask for permission
+     */
+    /*private void checkDownloadPermission() {
         if (ActivityCompat.shouldShowRequestPermissionRationale(this, android.Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
             Toast.makeText(this, "Нет, разрешения на загрузку файлов!", Toast.LENGTH_LONG).show();
         } else {
             ActivityCompat.requestPermissions(this, new String[]{android.Manifest.permission.WRITE_EXTERNAL_STORAGE}, 100);
         }
-        if (ActivityCompat.shouldShowRequestPermissionRationale(this, android.Manifest.permission.READ_EXTERNAL_STORAGE)) {
-            Toast.makeText(this, "Нет, разрешения на чтение файлов!", Toast.LENGTH_LONG).show();
-        } else {
-            ActivityCompat.requestPermissions(this, new String[]{android.Manifest.permission.READ_EXTERNAL_STORAGE}, 100);
-        }
-    }
+    }*/
+    //
     BroadcastReceiver attachmentDownloadCompleteReceive = new BroadcastReceiver() {
         @Override
         public void onReceive(Context context, Intent intent) {
@@ -92,6 +111,7 @@ public class XLSParser extends AppCompatActivity {
             }
         }
     };
+    // The above code is find the downloaded file and then opening it.
     private void FindFile(final long downloadId) {
         DownloadManager downloadManager = (DownloadManager) getSystemService(Context.DOWNLOAD_SERVICE);
         DownloadManager.Query query = new DownloadManager.Query();
@@ -101,38 +121,95 @@ public class XLSParser extends AppCompatActivity {
             @SuppressLint("Range") int downloadStatus = cursor.getInt(cursor.getColumnIndex(DownloadManager.COLUMN_STATUS));
             @SuppressLint("Range") String downloadLocalUri = cursor.getString(cursor.getColumnIndex(DownloadManager.COLUMN_LOCAL_URI));
             if ((downloadStatus == DownloadManager.STATUS_SUCCESSFUL) && downloadLocalUri != null) {
-                openDownloadedExcel(new File(Uri.parse(downloadLocalUri).getPath()));
+                downloaded = new File(Uri.parse(downloadLocalUri).getPath());
+                openDownloadedExcel(downloaded);
             }
         }
         cursor.close();
     }
     // delete the original file
     private void DeleteOld(String inputFile) {
-        file = new File(getExternalFilesDir(Environment.DIRECTORY_DOWNLOADS).toString() + File.separator + inputFile);
-        if (file!=null) file.delete();
+        if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            file = new File(getExternalFilesDir(Environment.DIRECTORY_DOWNLOADS).toString() + File.separator + inputFile);
+            if (file != null) file.delete();
+        }
         //Log.d("Moving", String.valueOf(succes));
     }
+    /**
+     * Reads the excel file and returns a list of items
+     *
+     * @param file The file that you want to read.
+     */
     private void openDownloadedExcel(File file) {
         if(file!=null) {
             try {
                 Items = ExcelUtils.readExcel(file);
-                ListView listView = findViewById(R.id.list);
-                ProgramAdapter adapter = new ProgramAdapter(this, Items.get(0), Items.get(1), Items.get(2));
-                listView.setAdapter(adapter);
+                Load();
                 Log.d("reader", Items.toString());
             } catch (IOException e) {
                 e.printStackTrace();
             }
         }
     }
+    /**
+     * This function opens the marks info fragment
+     *
+     * @param pos The position of the item in the list that you want to open.
+     */
+    public static void OpenInfo(int pos){
+        IsMarksInfo = true;
+        PositionSelected = pos;
+        FragmentTransaction ft = fm.beginTransaction();
+        ft.add(R.id.fragmentXLS, marksInfoFragment);
+        ft.commit();
+        FragmentTransaction ft2 = fm.beginTransaction();
+        ft2.hide(marksAllFragment);
+        ft2.commit();
+    }
+    /**
+     * This function is called when the user presses the back button.
+     * If the user is in the marksInfoFragment, then the user is taken back to the marksAllFragment.
+     * If the user is in the marksAllFragment, then the user is taken back to the main menu
+     */
+    @Override
+    public void onBackPressed() {
+
+        if (!IsMarksInfo) {
+            super.onBackPressed();
+
+        } else {
+
+            IsMarksInfo = false;
+            FragmentTransaction ft2 = fm.beginTransaction();
+            ft2.show(marksAllFragment);
+            ft2.commit();
+            FragmentTransaction ft = fm.beginTransaction();
+            ft.remove(marksInfoFragment);
+            ft.commit();
+        }
+
+    }
+    /**
+     * This function is used to display an error message to the user
+     *
+     * @param ErrorMessage The error message to display.
+     */
+    @Nullable
+    public static void PutErrorCode(String ErrorMessage){
+        if (ErrorMessage == null) ErrorMessage = "None";
+        FragmentTransaction ft = fm.beginTransaction();
+        ft.replace(R.id.fragmentXLS, new ErrorFragment());
+        ft.commit();
+        ErrorFragment.Error(ErrorMessage);
+    }
 }
  class ExcelUtils {
 
     /**
-     * Read Excel File
-     * @param file
-     * @throws FileNotFoundException
-     * @return
+     * Reads the contents of the Excel file and returns a list of lists
+     *
+     * @param file The file to be read.
+     * @return A list of lists.
      */
     public static List<List> readExcel(File file) throws IOException {
         if(file == null) {
@@ -142,7 +219,11 @@ public class XLSParser extends AppCompatActivity {
         List<List> All = new ArrayList<List>();
         List<String> Items = new ArrayList<String>();
         List<String> Vypiska = new ArrayList<String>();
-        List<String> Itog = new ArrayList<String>();
+        List<Double> Itog = new ArrayList<Double>();
+        List<String> NoShow = new ArrayList<String>();
+        List<String> Pass = new ArrayList<String>();
+        List<String> Disease = new ArrayList<String>();
+        List<String> Lateness = new ArrayList<String>();
         try {
             FileInputStream infile = new FileInputStream(file);
             Workbook workbook = new HSSFWorkbook(infile);
@@ -164,30 +245,66 @@ public class XLSParser extends AppCompatActivity {
                 //Read one line at a time
                 //Convert the contents of each grid to a string
                 String value = getCellAsString(row, 2, formulaEvaluator);
-                if(!value.equals(""))Vypiska.add(value);
+                Vypiska.add(value);
 
             }for (int r = 4; r<rowsCount; r++) {
                 Row row = sheet.getRow(r);
                 int cellsCount = row.getPhysicalNumberOfCells();
                 //Read one line at a time
                 //Convert the contents of each grid to a string
-                String value = getCellAsString(row, 3, formulaEvaluator);
-                if(!value.equals(""))Itog.add(value);
+                Double value = getCellAsDouble(row, 3, formulaEvaluator);
+                Itog.add(value);
+
+            }for (int r = 4; r<rowsCount; r++) {
+                Row row = sheet.getRow(r);
+                int cellsCount = row.getPhysicalNumberOfCells();
+                //Read one line at a time
+                //Convert the contents of each grid to a string
+                String value = ""+getCellAsInt(row, 4, formulaEvaluator);
+                NoShow.add(value);
+
+            }for (int r = 4; r<rowsCount; r++) {
+                Row row = sheet.getRow(r);
+                int cellsCount = row.getPhysicalNumberOfCells();
+                //Read one line at a time
+                //Convert the contents of each grid to a string
+                String value = ""+getCellAsInt(row, 5, formulaEvaluator);
+                Pass.add(value);
+
+            }for (int r = 4; r<rowsCount; r++) {
+                Row row = sheet.getRow(r);
+                int cellsCount = row.getPhysicalNumberOfCells();
+                //Read one line at a time
+                //Convert the contents of each grid to a string
+                String value = ""+getCellAsInt(row, 6, formulaEvaluator);
+                Disease.add(value);
+
+            }for (int r = 4; r<rowsCount; r++) {
+                Row row = sheet.getRow(r);
+                int cellsCount = row.getPhysicalNumberOfCells();
+                //Read one line at a time
+                //Convert the contents of each grid to a string
+                String value = ""+getCellAsInt(row, 7, formulaEvaluator);
+                Lateness.add(value);
 
             }
             All.add(Items);
             All.add(Vypiska);
             All.add(Itog);
+            All.add(NoShow);
+            All.add(Pass);
+            All.add(Disease);
+            All.add(Lateness);
             Log.d("reader", Items.toString());
 
             return All;
         } catch (Exception e) {
             /* proper exception handling to be here */
             Log.e("reader", e.toString());
+            if(e.equals(java.io.FileNotFoundException.class)) XLSParser.PutErrorCode("FILE_NOT_FOUND_PERMISSIONS?");
         }
         return All;
     }
-
     /**
      * Read the contents of each line in the excel file
      * @param row
@@ -205,7 +322,7 @@ public class XLSParser extends AppCompatActivity {
                     value = ""+cellValue.getBooleanValue();
                     break;
                 case Cell.CELL_TYPE_NUMERIC:
-                    double numericValue = cellValue.getNumberValue();
+                    float numericValue = (float) cellValue.getNumberValue();
                     if(HSSFDateUtil.isCellDateFormatted(cell)) {
                         double date = cellValue.getNumberValue();
                         SimpleDateFormat formatter = new SimpleDateFormat("dd/MM/yy");
@@ -226,6 +343,44 @@ public class XLSParser extends AppCompatActivity {
         }
         return value;
     }
+     static int getCellAsInt(Row row, int c, FormulaEvaluator formulaEvaluator){
+         int value = 0;
+         try {
+             Cell cell = row.getCell(c);
+             CellValue cellValue = formulaEvaluator.evaluate(cell);
+             switch (cellValue.getCellType()) {
+                 case Cell.CELL_TYPE_NUMERIC:
+                     int numericValue = (int) cellValue.getNumberValue();
+                         value = numericValue;
+                     break;
+                 default:
+                     break;
+             }
+         } catch (NullPointerException e) {
+             /* proper error handling should be here */
+             //Log.e("reader", e.toString());
+         }
+        return value;
+    }
+     static double getCellAsDouble(Row row, int c, FormulaEvaluator formulaEvaluator){
+         double value = 0.0;
+         try {
+             Cell cell = row.getCell(c);
+             CellValue cellValue = formulaEvaluator.evaluate(cell);
+             switch (cellValue.getCellType()) {
+                 case Cell.CELL_TYPE_NUMERIC:
+                     double numericValue = (double) cellValue.getNumberValue();
+                     value = numericValue;
+                     break;
+                 default:
+                     break;
+             }
+         } catch (NullPointerException e) {
+             /* proper error handling should be here */
+             //Log.e("reader", e.toString());
+         }
+         return value;
+     }
 
     /**
      * Simply determine whether an Excel file is an Excel file based on the type suffix name
